ChaosMod: Enable (almost) all warnings for GCC/Clang & treat all of them (#3708)
Also: 
- Decrease the warning level of MSVC to W2 and mark all warnings as errors for all compilers.
- Use this ooportunity to finally raise the cmake version.

+++ ChaosMod/Components/EffectDispatcher.h
@@ -116,8 +116,8 @@ class EffectDispatcher : public Component
 	void DispatchRandomEffect(DispatchEffectFlags dispatchEffectFlags = DispatchEffectFlag_None,
 	                          const std::string &suffix               = {});
 
-	void UpdateEffects(int deltaTime);
-	void UpdateMetaEffects(int deltaTime);
+	void UpdateEffects(float deltaTime);
+	void UpdateMetaEffects(float deltaTime);
 
 	void ClearEffect(const EffectIdentifier &effectId);
 	enum ClearEffectsFlags
